backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pillow==11.0.0
opencv-python-headless==4.10.0.84
numpy>=1.24,<2.1
scikit-image==0.24.0
google-generativeai==0.8.3
sqlalchemy==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
httpx==0.28.0
aiofiles==24.1.0
greenlet>=3.0.0
